+++ sys/include/alloc/kpalloc.h
@@ -2,7 +2,6 @@
 #define KPALLOC_H 1
 
 #include "limine.h"
-
 #include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
@@ -10,10 +9,10 @@
 // alloc/phys_allocator.c
 void        initkpalloc(const uint64_t offset,
                         const struct limine_memmap_response* memmap);
-uintptr_t   allocframe(size_t count);
-uintptr_t   allocmegaframe(size_t count);
-void        freeframe(uintptr_t frameptr, size_t count);
-void        freemegaframe(uintptr_t frameptr);
+uint64_t   allocframe(size_t count);
+uint64_t   allocmegaframe(size_t count);
+void        freeframe(uint64_t frameptr, size_t count);
+void        freemegaframe(uint64_t frameptr);
 uint64_t    hhdmoffset(void);
 
 // alloc/util.c
