+++ sys/src/alloc/conversions.c
@@ -2,10 +2,10 @@
 #include "alloc/kpalloc.h"
 #include "alloc/kvalloc.h"
 
-uintptr_t physical(void* virtual) {
+uint64_t physical(void* virtual) {
     return (uintptr_t)virtual - hhdmoffset();
 }
 
-void* virtual(uintptr_t physical) {
+void* virtual(uint64_t physical) {
     return (void*)(physical + hhdmoffset());
 }
