+++ sys/src/arch/x86_64/alloc/virt_allocator.c
@@ -9,8 +9,8 @@
 #include <stddef.h>
 #include <stdint.h>
 
-static uintptr_t kernelpgtbl = 0;
-uintptr_t readkernelpgtbl(void) {
+static uint64_t kernelpgtbl = 0;
+uint64_t readkernelpgtbl(void) {
     return kernelpgtbl;
 }
 
@@ -21,8 +21,8 @@ void initkvalloc(uint64_t physbase, uint64_t virtbase,
     for (size_t i = 0; i < memmap->entry_count; i++)
         if (memmap->entries[i]->type == LIMINE_MEMMAP_EXECUTABLE_AND_MODULES) {
             for (uint64_t j = 0; j < memmap->entries[i]->length; j += PAGE_SIZE) {
-                uintptr_t physaddr = memmap->entries[i]->base + j;
-                uintptr_t virtaddr = virtbase + (physaddr - physbase);
+                uint64_t physaddr = memmap->entries[i]->base + j;
+                uint64_t virtaddr = virtbase + (physaddr - physbase);
 
                 // todo: elf64 parser
                 mappage(new_l4t, virtaddr, physaddr, PAGE_PRESENT | PAGE_WRITABLE);
@@ -33,8 +33,8 @@ void initkvalloc(uint64_t physbase, uint64_t virtbase,
         if (memmap->entries[i]->type != LIMINE_MEMMAP_RESERVED &&
             memmap->entries[i]->type != LIMINE_MEMMAP_BAD_MEMORY)
             for (uint64_t j = 0; j < memmap->entries[i]->length; j += PAGE_SIZE) {
-                uintptr_t physaddr = memmap->entries[i]->base + j;
-                uintptr_t virtaddr = physaddr + hhdmoffset();
+                uint64_t physaddr = memmap->entries[i]->base + j;
+                uint64_t virtaddr = physaddr + hhdmoffset();
 
                 mappage(new_l4t, virtaddr, physaddr, PAGE_PRESENT | PAGE_WRITABLE | PAGE_NO_EXECUTE);
 
@@ -43,7 +43,7 @@ void initkvalloc(uint64_t physbase, uint64_t virtbase,
                 }
             }
 
-    uintptr_t stacks = allocmegaframe(1);
+    uint64_t stacks = allocmegaframe(1);
     mappage(new_l4t, STACK_BASE_VIRT, stacks, PAGE_PRESENT | PAGE_WRITABLE | PAGE_HUGE | PAGE_NO_EXECUTE);
 
     kernelpgtbl = physical(new_l4t);
@@ -51,19 +51,19 @@ void initkvalloc(uint64_t physbase, uint64_t virtbase,
 }
 
 void mappage(pagetable_t* globaltbl,
-                    uintptr_t va, uintptr_t pa, uint64_t flags) {
+                    uint64_t va, uint64_t pa, uint64_t flags) {
     uint64_t index;
 
     index = LVL4_INDEX(va);
     if (!(globaltbl->entries[index] & PAGE_PRESENT)) {
-        uintptr_t lowertbl = allocframe(1);
+        uint64_t lowertbl = allocframe(1);
         globaltbl->entries[index] = lowertbl | PAGE_PRESENT | PAGE_WRITABLE;
     }
 
     pagetable_t* uppertbl = virtual(globaltbl->entries[index] & PTE_ADDRESS_MASK);
     index = LVL3_INDEX(va);
     if (!(uppertbl->entries[index] & PAGE_PRESENT)) {
-        uintptr_t lowertbl = allocframe(1);
+        uint64_t lowertbl = allocframe(1);
         uppertbl->entries[index] = lowertbl | PAGE_PRESENT | PAGE_WRITABLE;
     }
 
@@ -81,7 +81,7 @@ void mappage(pagetable_t* globaltbl,
         panic("4kib page allocation collided with existing 2mib page.");
 
     if (!(middletbl->entries[index] & PAGE_PRESENT)) {
-        uintptr_t lowertbl = allocframe(1);
+        uint64_t lowertbl = allocframe(1);
         middletbl->entries[index] = lowertbl | PAGE_PRESENT | PAGE_WRITABLE;
     }
